% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{ACheatmap}
\alias{ACheatmap}
\title{Heatmap of a matrix.}
\usage{
ACheatmap(
  Mat,
  center_value = 0,
  col.upper = "darkred",
  col.center = "grey95",
  col.lower = "#3B9AB2",
  col.n_breaks = 59,
  use_x11_device = TRUE,
  remove_diag = FALSE,
  main = " ",
  x_label = " ",
  y_label = " ",
  horizontal = TRUE
)
}
\arguments{
\item{Mat}{the matrix to be plotted.}

\item{center_value}{the value around which the palette has to be centered. Set NULL for symmetric palette.}

\item{col.upper}{the color for the highest value in the matrix.}

\item{col.center}{the color for the centered value (if desired) or for the middle value in the matrix (oterhwise).}

\item{col.lower}{the color for the lowest value in the matrix.}

\item{col.n_breaks}{has to be odd. The refinement of the palette.}

\item{use_x11_device}{boolean, if x11() device has to be activated or not.}

\item{remove_diag}{boolean, if the diagonal has to be removed or not.}

\item{main}{the title of the plot.}

\item{x_label}{the name of the x-axis in the plot.}

\item{y_label}{the name of the y-axis in the plot.}

\item{horizontal}{boolean, if the heatmap bar has to be horizontal or not.}
}
\value{
this function does not return anything
}
\description{
It plots the heatmap of a given matrix. It is usefull because it allows the user to use custom, non-symmetric palettes, centered around a desired value.
}
