% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{BFDR_selection}
\alias{BFDR_selection}
\title{Bayesian FDR Analysis}
\usage{
BFDR_selection(
  plinks,
  tol = seq(0.1, 1, by = 0.025),
  min_rate = 0.05,
  diag = F
)
}
\arguments{
\item{plinks}{matrix of size \eqn{p x p} containing the posterior inclusion probability for each link. It has to be upper triangular.}

\item{tol}{sequence of tolerances to be tested trying to select a graph truncating \code{plinks} at that value.}

\item{min_rate}{fix false discoveries to remain under this selected threshold.}

\item{diag}{boolean, if the diagonal of \code{plinks} has to be included in the computations. Set \code{FALSE} if the graph is in complete form, set \code{TRUE} for block graphs.}
}
\value{
a list of two elements: best_threshold, the best value of tol according to this analysis.
best_truncated_graph, the proposed posterior graph according to the analysis.
}
\description{
Given the plinks matrix, this utility computes the False Discovery Rate Index, forcing the false discovery rates to be less than \code{min_rate.}
}
