% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{Block2Complete}
\alias{Block2Complete}
\title{Block to Complete map}
\usage{
Block2Complete(Gblock, groups)
}
\arguments{
\item{Gblock}{matrix of size \eqn{n_groups x n_groups} to be mapped.}

\item{groups}{list representing the groups of the block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. It is possible to generete it through CreateGroups function.}
}
\value{
p x p matrix containing the Complete form of Gblock.
}
\description{
Function for mapping from a block matrix to its non-block representation. It can be used also for matrices whose entries are not just 1 or 0.
}
