% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{Compute_AUC}
\alias{Compute_AUC}
\title{Compute AUC Index}
\usage{
Compute_AUC(x, y)
}
\arguments{
\item{x}{Numeric vector of length \code{n}.}

\item{y}{Numeric vector of length \code{n}.}
}
\value{
A numeric value which is the area under the ROC curve.
}
\description{
Function to compute the area under ROC curve exploiting trapezoidal rule.
}
