% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Compute_Quantiles}
\alias{Compute_Quantiles}
\title{Compute quantiles of sampled values}
\usage{
Compute_Quantiles(
  file_name,
  p,
  n,
  stored_iterG,
  stored_iter = 0L,
  Beta = FALSE,
  Mu = FALSE,
  TauEps = FALSE,
  Precision = FALSE,
  prec_elem = 0L,
  lower_qtl = 0.05,
  upper_qtl = 0.95
)
}
\description{
Compute quantiles of sampled values
}
