% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CreateGroups}
\alias{CreateGroups}
\title{Create Groups}
\usage{
CreateGroups(p, n_groups)
}
\description{
This function creates a list with the groups. If possible, groups of equal size are created. The goal of this function is to fix a precise notation that will be used in all the code.
It is indeed recommended to use this function to create them as they need to follow a precise notation.
}
