% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Extract_Chain}
\alias{Extract_Chain}
\title{Extract chain from file}
\usage{
Extract_Chain(file_name, variable, stored_iter, n, p, index1, index2 = 0L)
}
\description{
This function extract the sampled chain for the \code{index1}-th component of variable specified in \code{variable}. For \code{"Beta"} coefficients one has to use also \code{index2}.
}
