% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FGM_sampling_c}
\alias{FGM_sampling_c}
\title{Functional Graphical model for smoothing}
\usage{
FGM_sampling_c(
  data,
  niter,
  burnin,
  thin,
  thinG,
  BaseMat,
  Beta0,
  mu0,
  tau_eps0,
  G0,
  K0,
  a_tau_eps,
  b_tau_eps,
  sigmamu,
  bK,
  DK,
  sigmaG,
  paddrm,
  Gprior,
  MCprior,
  MCpost,
  threshold,
  form = "Complete",
  prior = "Uniform",
  algo = "MH",
  groups = NULL,
  seed = 0L,
  print_info = TRUE
)
}
\description{
Functional Graphical model for smoothing
}
