% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{FLM_sampling}
\alias{FLM_sampling}
\title{Functional Linear model for smoothing}
\usage{
FLM_sampling(
  p,
  data,
  niter = 1e+05,
  burnin = niter/2,
  thin = 1,
  diagonal_graph = T,
  G = NULL,
  Param = NULL,
  HyParam = NULL,
  Init = NULL,
  progressBar = TRUE,
  seed = 0
)
}
\arguments{
\item{p}{number of basis functions.}

\item{data}{matrix of dimension \mjseqn{n_grid_points \times n} containing the evaluation of \code{n} functional data over a grid of \code{n_grid_points} nodes.}

\item{niter}{the number of total iterations to be performed in the sampling. The number of saved iteration will be \code{(niter - burnin)/thin}.}

\item{burnin}{the number of discarded iterations.}

\item{thin}{the thining value, it means that only one out of \code{thin} itarations is saved.}

\item{diagonal_graph}{boolean, set true if the graph has to be diagonal. Set false otherwise and pass as input the desired graph through the \code{G} parameter.}

\item{G}{matrix of size \code{p x p} representing the graphical part of the model that would remain fixed through out the sampling. Needed only if \code{diagonal_graph} is set to \code{FALSE}.}

\item{Param}{list containing parameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{sampler_parameters}} function.
It is indeed recommended to build it through that particular function. It is important to remember that \code{BaseMat} field is needed and cannot be defaulted. Use \code{\link{Generate_Basis}} to create it.
It has to be a matrix of dimension \mjseqn{n_grid_points \times p} containing the evalutation of \code{p} Bspline basis over a grid of \code{n_grid_points} nodes.}

\item{HyParam}{list containing hyperparameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{LM_hyperparameters}} function. It is indeed recommended to build it through that particular function.}

\item{Init}{list containig initial values for Markov chain. It has to follow the same notation of the one generated by \code{\link{LM_init}} function. It is indeed recommended to build it through that particular function.}

\item{progressBar}{boolean, set \code{TRUE} to display the progress bar.}

\item{seed}{integer, seeding value. Set 0 for random seed.}

\item{BaseMat}{matrix of dimension \code{n_grid_points x p} containing the evalutation of \code{p} Bspline basis over a grid of \code{n_grid_points} nodes.}
}
\value{
Two lists are returned, one with all the sampler values and the other with the posterior means.
}
\description{
\loadmathjax This function performs a linear regression for functional data, according to model (INSERIRE FORMULA MODELLO).
It is not a graphical model, the graph has to be fixed. It is possible to fix both a diagonal graph or a generic graph.
}
