% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FLM_sampling_c}
\alias{FLM_sampling_c}
\title{Functional Linear model for smoothing}
\usage{
FLM_sampling_c(
  data,
  niter,
  burnin,
  thin,
  BaseMat,
  G,
  Beta0,
  mu0,
  tau_eps0,
  tauK0,
  K0,
  a_tau_eps,
  b_tau_eps,
  sigmamu,
  aTauK,
  bTauK,
  bK,
  DK,
  file_name,
  diagonal_graph = TRUE,
  threshold_GWish = 1e-08,
  seed = 0L,
  print_info = TRUE
)
}
\arguments{
\item{data}{matrix of dimension r x n containing the evaluation of n functional data over a grid of r nodes.}

\item{niter}{the number of total iterations to be performed in the sampling. The number of saved iteration will be (niter - burnin)/thin.}

\item{burnin}{the number of discarded iterations.}

\item{thin}{the thining value, it means that only one out of thin itarations is saved.}

\item{BaseMat}{matrix of dimension r x p containing the evalutation of p Bspline basis over a grid of r nodes}
}
\description{
This function performs a linear regression for functional data, according to model (INSERIRE FORMULA MODELLO).
It is not a graphical model, indeed the precision matrix for the regression coefficients is chosen diagonal.
}
