% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Generate_Basis_derivatives}
\alias{Generate_Basis_derivatives}
\title{Generate Bspine basis and its derivatives}
\usage{
Generate_Basis_derivatives(
  n_basis,
  nderiv,
  range,
  n_points = 0L,
  grid_points = as.numeric(c(0)),
  order = 3L
)
}
\arguments{
\item{n_basis}{number of basis functions.}

\item{nderiv}{number of derivates that have to be computed.}

\item{range}{vector of two elements containing first the lower and then the upper bound of the interval.}

\item{grid_points}{vector of points where the basis has to be evaluated. If defaulted, then n_points are uniformly generated in the interval.}

\item{order}{order of the Bsplines. Set four for cubic splines.}

\item{points}{number of grid points where the basis has to be evaluated. It is not used if the points are directly passed in the grid_points parameter.}
}
\value{
a List of length nderiv+1 such that each element is a grid_points.size() x n_basis matrix representing the evaluation of
all the k-th derivative of all the splines in all the grid points.
}
\description{
This function creates a truncated Bspline basis in the interval \link{range(1),range(2)} and evaluate them over a grid of points up to derivative of order nderiv.
For convention, derivatives of order 0 are the splines themselves. This implimes that the first returned element is always equal to the output of the function Generate_Basis.
}
