% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GGM_sampling}
\alias{GGM_sampling}
\title{Sampler for Guassian Graphical Models}
\usage{
GGM_sampling(
  data,
  n,
  niter = 1e+05,
  burnin = niter/2,
  thin = 1,
  Param = NULL,
  HyParam = NULL,
  Init = NULL,
  file_name = "GGMresults",
  form = "Complete",
  prior = "Uniform",
  algo = "RJ",
  groups = NULL,
  n_groups = 0,
  seed = 0,
  print_info = TRUE
)
}
\arguments{
\item{data}{two possibilities are available. (1) a \mjseqn{p \times n} matrix corresponding to \code{n} observation of \code{p}-dimensional variables or (2)
a \mjseqn{p \times p} matrix representing \mjseqn{\sum_{i=1}^{n}(y_{i}y_{i}^{T})}, where \mjseqn{y_{i}} is the \code{i}-th observation of a \code{p}-dimensional variale.
Data are assumed to be zero mean.}

\item{n}{integer, number of observed data.}

\item{niter}{the number of total iterations to be performed in the sampling. The number of saved iteration will be \mjseqn{(niter - burnin)/thin}.}

\item{burnin}{number of discarded iterations.}

\item{thin}{the thining value, it means that only one out of \code{thin} itarations is saved.}

\item{Param}{list containing parameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{sampler_parameters}} function. It is indeed recommended to build it through that particular function. \code{BaseMat} field is not needed.}

\item{HyParam}{list containing hyperparameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{GM_hyperparameters}} function. It is indeed recommended to build it through that particular function.}

\item{Init}{list containig initial values for Markov chain. It has to follow the same notation of the one generated by \code{\link{GM_init}} function. It is indeed recommended to build it through that particular function.}

\item{form}{string that states if the true graph has to be in \code{"Block"} or \code{"Complete"} form. Only possibilities are \code{"Complete"} and \code{"Block"}.}

\item{prior}{string with the desidered prior for the graph. Possibilities are \code{"Uniform"}, \code{"Bernoulli"} and for \code{"Block"} graphs only \code{"TruncatedBernoulli"} and \code{"TruncatedUniform"} are also available.}

\item{algo}{string with the desidered algorithm for sampling from a GGM. Possibilities are \code{"MH"}, \code{"RJ"} and \code{"DRJ"}.}

\item{groups}{a list representing the groups of the block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. If \code{"NULL"}, \code{"n_groups"} are automatically generated. Not needed if form is set to \code{"Complete"}.}

\item{n_groups}{number of desired groups. Not used if form is \code{"Complete"} or if the groups are directly insered as \code{groups} parameter.}

\item{seed}{integer, seeding value. Set 0 for random seed.}

\item{print_info}{boolean, if true progress bar and execution time are displayed.}
}
\value{
It returns a list composed of: \code{MeanK}, the posterior mean of all sampled precision matrix, \code{plinks} which contains the posterior probability of inclusion of each possible link.
It is a \mjseqn{p \times p} matrix if \code{form} is \code{"Complete"}, or a \mjseqn{n\_groups \times n\_groups} matrix if \code{form} is \code{"Block"}. \code{AcceptedMoves} contains the number of
Metropolis-Hastings moves that were accepted in the sampling, \code{VisitedGraphs} the number of graph that were visited at least once, \code{TracePlot_Gsize} is a vector of length \mjseqn{(niter - burnin)/thin}
such that each element is equal to the size of the visited graph in that particular iteration and finally \code{SampledGraphs} is a list containing all the visited graphs and their absolute frequence of visit.
To save memory, the graphs are represented only by the upper triangular part, stored row-wise.
A binary \code{".h5"} file named \code{file_name} is also generated, it contains all the sampled values.
}
\description{
\loadmathjax This function draws samples a posteriori from a Gaussian Graphical Models. The prior chosen for the precision matrix is a GWishart, whose parameters can be fixed in \code{HyParam}.
Diffefent priors are available for the graph, they can be set via \code{prior} input.
}
