% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GGM_sampling_c}
\alias{GGM_sampling_c}
\title{Sampler for Guassian Graphical Models}
\usage{
GGM_sampling_c(
  data,
  p,
  n,
  niter,
  burnin,
  thin,
  D,
  b,
  G0,
  K0,
  MCprior = 100L,
  MCpost = 100L,
  threshold = 1e-08,
  form = "Complete",
  prior = "Uniform",
  algo = "MH",
  groups = NULL,
  seed = 0L,
  Gprior = 0.5,
  sigmaG = 0.1,
  paddrm = 0.5,
  print_info = TRUE
)
}
\arguments{
\item{data}{matrix of size \eqn{p x p} containing \eqn{\sum(Y_i^{T}Y_i)}. Data are required to be zero mean.}

\item{p}{non necessario}

\item{n}{number of observed data.}

\item{niter}{number of total iterations to be performed in the sampling. The number of saved iteration will be \eqn{(niter - burnin)/thin}.}

\item{burnin}{number of discarded iterations.}

\item{thin}{thining value, it means that only one out of thin itarations is saved.}

\item{D}{matrix of double stored columnwise. It is prior GWishart inverse scale parameter. It has to be symmetric and positive definite.}

\item{b}{double, it is prior GWishart shape parameter. It has to be larger than 2 in order to have a well defined distribution.}

\item{MCprior}{positive integer, the number of iteration for the MonteCarlo approximation of posterior normalizing constant of GWishart distribution. Needed only if algo is set to \code{"MH"}.}

\item{threshold}{double, threshold for convergence in GWishart sampler.}

\item{form}{string that may take as values only \code{"Complete"} of \code{"Block"}. It states if the algorithm has to run with \code{"Block"} or \code{"Complete"} graphs.}

\item{prior}{string with the desidered prior for the graph. Possibilities are \code{"Uniform"}, \code{"Bernoulli"} and for \code{"Block"} graphs only \code{"TruncatedBernoulli"} and \code{"TruncatedUniform"} are also available.}

\item{algo}{string with the desidered algorithm for sampling from a GGM. Possibilities are \code{"MH"}, \code{"RJ"} and \code{"DRJ"}.}

\item{seed}{set 0 for random seed.}

\item{Gprior}{double representing the prior probability of inclusion of each link in case \code{"Bernoulli"} prior is selected for the graph. Set 0.5 for uniform prior.}

\item{sigmaG}{double, the standard deviation used to perturb elements of precision matrix when constructing the new proposed matrix.}

\item{paddrm}{double, probability of proposing a new graph by adding one link.}

\item{print_info}{boolean, if \code{TRUE} progress bar and execution time are displayed.}

\item{n_groups}{int, number of desired groups. Not used if form is \code{"Complete"} or if the groups are directly insered as group parameter. (CHE PER ORA NON ESISTE)}
}
\value{
This function returns a list with the posterior precision mean, a matrix with the probability of inclusion of each link, the number of accepted moves, the number of visited graphs and the list of all visited graphs.
}
\description{
This function draws samples a posteriori from a Gaussian Graphical Models. NON MI SERVE ESPORTARLA
}
