% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GM_hyperparameters}
\alias{GM_hyperparameters}
\title{Skeleton for Hyperparameters in Graphical Models}
\usage{
GM_hyperparameters(
  a_tau_eps = 20,
  b_tau_eps = 0.002,
  sigma_mu = 100,
  b_K = 3,
  D_K = NULL,
  p = NULL,
  p_addrm = 0.5,
  sigmaG = 0.5,
  Gprior = 0.5
)
}
\arguments{
\item{a_tau_eps}{double, shape parameter for gamma prior distribution of \code{tau_eps} parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{b_tau_eps}{double, inverse scale parameter for gamma prior distribution of \code{tau_eps} parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{sigma_mu}{double, covariance for normal multivariate distribution of mu parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{b_K}{double, it is prior GWishart shape parameter. It has to be larger than 2 in order to have a well defined distribution. Needed in all graphical model, i.e \code{FGM} and \code{GGM} sampler and in \code{FLM} with fixed graph.}

\item{D_K}{matrix of double stored columnwise. It is prior GWishart inverse scale parameter. It has to be symmetric and positive definite. Default is identity matrix but \code{p} has to be provided in that case. Needed in all graphical model, i.e \code{FGM} and \code{GGM} sampler and in \code{FLM} with fixed graph.}

\item{p}{integer, dimension of the graph. It is needed if \code{D_K} parameter is null.}

\item{p_addrm}{double, probability of proposing a new graph by adding one link. Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler.}

\item{sigmaG}{double, the standard deviation used to perturb elements of precision matrix when constructing the new proposed matrix. Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler. If algorithm is \code{"MH"} it is not used.}

\item{Gprior}{double representing the prior probability of inclusion of each link in case \code{"Bernoulli"} prior is selected for the graph. Set 0.5 for \code{"Uniform"} prior. Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler.}
}
\value{
A list with all the inserted hyperparameters.
}
\description{
This function simply creates a skeleton for all the hyperparameters that has to be fixed in graphical samplers. All the quantities has a default value,
moreover some of them may not be needed in some algorithm. The goal of this function is to fix a precise notation that will be used in all the code.
}
