% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{Graph_distance}
\alias{Graph_distance}
\title{Graph_distance}
\usage{
Graph_distance(G1, G2, diag = F)
}
\arguments{
\item{G1}{first graph}

\item{G2}{second graph}

\item{diag}{if diagonal should be included or not}
}
\description{
Graph_distance
}
