% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GraphTest}
\alias{GraphTest}
\title{Function for testing properties of Graph classes}
\usage{
GraphTest()
}
\arguments{
\item{No}{parameters are required}
}
\description{
Function for testing properties of Graph classes
}
