% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GWish_PriorConst_tuning}
\alias{GWish_PriorConst_tuning}
\title{Tuning MC iterations for prior normalizing constant of GWishart distribution}
\usage{
GWish_PriorConst_tuning(
  p,
  b = 3,
  D = diag(p),
  nrep = 100,
  MCiter_list = c(100, 250, 500, 1000, 2500, 5000, 10000)
)
}
\arguments{
\item{p}{The dimension of the graph in complete form}

\item{b}{Shape parameter. It has to be larger than 2 in order to have a well defined distribution.}

\item{D}{Inverse scale matrix. It has to be symmetric and positive definite.}

\item{nrep}{How many repetitions has to be performed for each possible threshold}

\item{MCiter_list}{The list containg the Monte Carlo iterations to be taken into consideration}
}
\value{
A list with four possible MCiteration levels
}
\description{
This function selects the number of MonteCarlo iterations when computing a the GWishart prior normalizing constant. The problem is that the MonteCarlo
method by ATACK-MASSAM requires computing exp(-0.5 sum(Phi_ij)^2). As the number of nodes in the graph grows, that quantity is smaller and smaller and
it is possible that it is lower than zero machina. If that is the case, the number of MC iterarions needs to be larger.
This function propose some possible MC iterations loop according to the percentage of computed constant different from -Inf out of nrep trials.
Four possibilities are returned according to four possible levels, that are 0.95, 0.98, 0.99 and 1. If some level is not reach in the selected threshold level, then NA is returned
}
