% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GWish_sampler_tuning}
\alias{GWish_sampler_tuning}
\title{Tuning the threshold for the GWishart sampler}
\usage{
GWish_sampler_tuning(
  p,
  nrep = 1000,
  initial_threshold = 0.01,
  max_threshold = 1e-14,
  norm = "Mean"
)
}
\arguments{
\item{p}{The dimension of the graph in complete form}

\item{nrep}{How many repetitions has to be performed for each possible threshold}

\item{initial_threshold}{The first and larger threshold to be tested}

\item{max_threshold}{The last and smaller threshold to be tested}

\item{norm}{The norm with respect to which convergence happens}
}
\value{
A list with four possible threshold levels and the exepected number of iteration for each threshold
}
\description{
This function select possible threshold for the convergence of the matrix in the GWishart sampler. The sampler starts drawing a sampler from a Wishart
and then starts a loop for imposing all the zeros in the right positions. The loop has to be stopped according to some convergence criteria and this
that not always the structure of the graph is pefectly respected. This function propose some threshold for stopping the loop according to the percentage
of sampled matrix that have the right structure out of nrep trials.
Four possibilities are returned according to four possible levels, that are 0.95, 0.98, 0.99 and 1. If some level is not reach in the selected threshold level, then NA is returned
}
