% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{LM_init}
\alias{LM_init}
\title{Skeleton for initial values in Linear model}
\usage{
LM_init(
  p,
  n,
  empty = TRUE,
  K0 = NULL,
  Beta0 = NULL,
  mu0 = NULL,
  tau_eps0 = NULL,
  tauK0 = NULL,
  seed = 0
)
}
\arguments{
\item{p}{integer, dimension of the graph.}

\item{n}{integer, number of observed data.}

\item{empty}{boolean, if \code{TRUE} all chians start from zero. If not, it is possible to set the desidered value explicitely by means of the other function parameters or just leave them \code{NULL} and random initial points will be used.}

\item{K0}{numeric \mjseqn{p \times p} matrix representing the initial precision matrix. It will not actually be used if \code{\link{FLM_sampling}} is called with \code{diagonal_graph} set to \code{TRUE}.
If \code{NULL}, a random matrix is selected.}

\item{Beta0}{numeric \mjseqn{p \times n} matrix representing the initial beta coefficients. If \code{NULL}, they are randomly generated.}

\item{mu0}{numeric \mjseqn{p}-dimensional vector representing the initial mu coefficients. If \code{NULL}, they are randomly generated.}

\item{tau_eps0}{scalar number representing the initial tau_eps coefficients. If \code{NULL}, it is randomly generated.}

\item{tauK0}{numeric \mjseqn{p}-dimensional vector representing the initial values for the diagonal of precision matrix. It will not actually be used if \code{\link{FLM_sampling}} is called with \code{diagonal_graph} set to \code{FALSE}.
If \code{NULL}, they are randomly generated.}

\item{seed}{integer, seeding value. Set 0 for random seed.}
}
\value{

}
\description{
\loadmathjax This function simply creates a structure for setting initial values for Markov chians in \code{\link{FLM_sampling}}. All the quantities has a default value,
that can be set to empty chains or random initial points. Some quantities are not needed, it depends on what version of \code{\link{FLM_sampling}} is called. However they are all initialized.
The goal of this function is to fix a precise notation that will be used in all the code.
}
