% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purees.R
\docType{data}
\name{purees}
\alias{purees}
\title{Absorbance mid-IR spectra of 351 strawberry fruit purées.}
\format{
A list that contains two elements. \strong{wavelengths} contains the above-mentioned
vector of wavelengths, while \strong{data} is a data frame with 351 rows and 235 variables:
}
\source{
\url{https://csr.quadram.ac.uk/example-datasets-for-download/}
}
\usage{
data("purees")
}
\description{
This dataset contains the absorbance spectra of strawberry fruit purées, evaluated
in 235 different wavelenghts.
}
\keyword{datasets}
