% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read_InfoFile}
\alias{Read_InfoFile}
\title{Read information from file}
\usage{
Read_InfoFile(file_name)
}
\arguments{
\item{file_name, }{string with the name of the file to be open. It has to include the extension, usually \code{.h5}.}
}
\value{
It returns a list containing \code{p}, the dimension of the graph, \code{n} the number of observed data, \code{stored_iter} the number of saved iterations for the regression parameters,
\code{stored_iterG} the number of saved iterations for the graphical related quantities, i.e the graph and the precision matrix.
}
\description{
\loadmathjax Read from \code{file_name} some information that are needed to extract data from it.
}
