% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnormal}
\alias{rmvnormal}
\title{Sampler for Multivariate Normal random variables}
\usage{
rmvnormal(mean, Mat, isPrec = FALSE, isChol = FALSE, isUpper = FALSE)
}
\arguments{
\item{mean}{Eigen column vector of size p representig the mean of the Gaussian distribution.}

\item{Mat}{Eigen p x p column major matrix reprenting the covariance or the precision or their Cholesky decompositions.}

\item{isPrec}{bool, set true if Mat parameter is a precision, false if it is a covariance.}

\item{isChol}{bool, set true if Mat parameter is a triangular matrix representig the Cholesky decomposition of the precision or covariance}

\item{isUpper}{bool, used only if isChol is true. Set true if Mat is upper triangular, false if lower.}
}
\value{
It returns a p dimensional vector with the sampled values.
}
\description{
This function draws random samples from Multivariate Gaussian distribution. It implements both covariance and precision parametrization.
It is also possible to pass directly the Cholesky decomposition if it is available before the call.
}
