% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rnormal}
\alias{rnormal}
\title{Sampler for Normal distribution}
\usage{
rnormal(mean = 0, sd = 1, seed = 0L)
}
\arguments{
\item{mean}{the mean value.}

\item{sd}{the standard deviation.}

\item{seed}{integer, seeding value. Set 0 for random seed.}
}
\value{
A single drawn values from N(mean,sd*sd).
}
\description{
This function draw a sample from the Gaussian distribution.
}
