% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{Sensitivity_analysis}
\alias{Sensitivity_analysis}
\title{Bayesian Sensitivity Analysis}
\usage{
Sensitivity_analysis(
  PL,
  G_true,
  tol = seq(0.1, 1, by = 0.01),
  ROC = FALSE,
  diag = F
)
}
\arguments{
\item{PL}{matrix of size \eqn{p \times p} containing the posterior inclusion probability for each link. It has to be upper triangular.}

\item{G_true}{matrix of size \eqn{p \times p} containing the true graph. It has to be upper triangular.}

\item{tol}{sequence of tolerances to be tested trying to select a graph truncating \code{plinks} at that value.}

\item{ROC}{boolean. If \code{TRUE}, the plot the ROC curve is showed.}

\item{diag}{}
}
\value{
A list of 5 elements: all_confusion, a list with all the confusion matrices, one for each tol value. best_threshold, the best value of tol according to this analysis.
best_confusion, the best confusion matrix, the one corresponding to best_threshold. best_truncated_graph, the proposed posterior graph according to the analysis.
AUC, the AUC Index corresponding to the best value selected.
}
\description{
Given the true graph and a \code{plinks} matrix, this utility computes the confusion matrices, plot the ROC curve (if required) and the AUC index,
selects the best threshold according to the number of misclassified links and also return the best graph according to the abovementioned analysis.
}
