% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{simulate_curves}
\alias{simulate_curves}
\title{Simulate curves}
\usage{
simulate_curves(
  p = 10,
  n = 300,
  r = 235,
  range_x = c(100, 200),
  G = NULL,
  K = NULL,
  b = 3,
  D = NULL,
  tau_eps = 0,
  rate = 0.01/2,
  spline_order = 3,
  n_picks = 1,
  height1 = 1,
  height2 = 1,
  width1 = 36,
  width2 = 24,
  position1 = 10,
  position2 = 2,
  n_plot = n,
  highlight_band1 = NULL,
  highlight_band2 = NULL,
  title_plot = "Curves",
  xtitle = " ",
  ytitle = " ",
  seed = 1212
)
}
\arguments{
\item{p}{integer, the number of basis functions. It also represents the dimension of the true underlying graph.}

\item{n}{integer, the number of curves to be generated.}

\item{r}{integer, the dimension of the grid where the spline are evaluated.}

\item{range_x}{the range where the curves has to be plotted. Not needed if \code{n_plot} is 0.}

\item{G}{matrix of size \mjseqn{p \times p} representing the true underlying graph. Only the upper triangular part is needed. Default is \code{NULL} that corresponds to a diagonal graph.}

\item{K}{matrix of size \mjseqn{p \times p} representing the true underlying precision matrix. If \code{NULL}, a sample from a GWishart(\code{b}, \code{D}) is drawn.}

\item{b}{GWishart Shape parameter used if the matrix \code{K} has to be drawn. It has to be larger than 2 in order to have a well defined distribution.}

\item{D}{GWishart Inverse Scale matrix parameter used if the matrix \code{K} has to be drawn. It has to be \mjseqn{p \times p}, symmetric and positive definite.}

\item{tau_eps}{numerical value, it represents the precision of the Gaussian noise to be added to the curves. Set 0 to have smooth curve, no noise is added in that case.}

\item{rate}{If \code{G} is \code{NULL}, only the diagonal elements of \code{K} are generated. They are sampled from a \mjseqn{Gamma(Shape = 5, Rate = rate)}.}

\item{spline_order}{order of the Bsplines. Set four for cubic splines.}

\item{n_picks}{The number of desired picks in the simulated curves. It may be one for a single, Gaussian like, shape or two for a double Gaussian shape.}

\item{height1}{Heigth of the pick of the first Gaussian.}

\item{height2}{Heigth of the pick of the second Gaussian.}

\item{width1}{Set how large the first pick has to be. This parameter acts like a standard deviation of a Normal distribution, so the larger it is, the thinner is the pick.}

\item{width2}{Set how large the second pick has to be. This parameter acts like a standard deviation of a Normal distribution, so the larger it is, the thinner is the pick.}

\item{position1}{Set where the first pick has to be located. Position is then computed as \mjseqn{range\_x\[1\] + (range\_x\[2\]-range\_x\[1\])/position1}.}

\item{position2}{Set where the second pick has to be located. Position is then computed as \mjseqn{range\_x\[2\] - (range\_x\[2\]-range\_x\[1\])/position2}. Set 2 to place it in the middle.}

\item{n_plot}{the number of curves to be plotted. Set 0 for no plot.}

\item{highlight_band1}{a vector that states if a particular band of the plot has to be highlighted. It has to be a vector within the range \mjseqn{\[1,n\_basis\]}.}

\item{highlight_band2}{a vector that states if a particular band of the plot has to be highlighted. It has to be a vector within the range \mjseqn{\[1,n\_basis\]}.}

\item{title_plot}{the title of the plot.}

\item{xtitle}{the title of the x-axis.}

\item{ytitle}{the title of the x-axis.}

\item{seed}{seed used to simulate data. Set 0 for random seed.}
}
\value{
It returns a list composed of: all simulated curves in \code{data}, all underlying parameters, called \code{Beta}, \code{mu}, \code{K} and the true graph \code{G}.
It also return the design matrix \code{basemat}, generated by \code{\link{Generate_Basis}} and the vector of the internal knots generating the splines, see \code{\link{Generate_Basis}}.
}
\description{
\loadmathjax This function genrates a dataset of n functional curves. Data are generated trying to simulate the shape of one or two Gaussian distributions.
}
