% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{smooth_curves_credible_bands}
\alias{smooth_curves_credible_bands}
\title{Computes and plot smoothed curves with credible bands}
\usage{
smooth_curves_credible_bands(
  beta,
  betaLower,
  betaUpper,
  BaseMat,
  n_plot = 0,
  range = NULL,
  grid_points = NULL,
  internal_knots = NULL,
  highlight_band1 = NULL,
  highlight_band2 = NULL,
  title_plot = "Smoothed Curves",
  xtitle = " ",
  ytitle = " "
)
}
\arguments{
\item{beta}{matrix of dimension n_basis x n_curves containing the mean values of regression coefficients}

\item{betaLower}{matrix of dimension n_basis x n_curves containing the lower quantiles values of regression coefficients. Can be obtained by Compute_QuantileBeta function.}

\item{betaUpper}{matrix of dimension n_basis x n_curves containing the upper quantiles values of regression coefficients. Can be obtained by Compute_QuantileBeta function.}

\item{BaseMat}{matrix of dimension n_grid_points x n_basis containing the evaluation of all the spline in all the grid points}

\item{n_plot}{the number of curves to be plotted. Set 0 for no plot.}

\item{range}{the range where the curves has to be plotted. Not needed if n_plot is 0.}

\item{grid_points}{vector of size n_grid_points with the points where the splines are evaluated. If defaulted they are uniformly generated. Not needed if n_plot is 0.}

\item{internal_knots}{vector with the internal knots used to construct the splines. Can be obtained as return value of function Generate_Basis. Default is null.
but if provided, n_basis are displayed in the plot. The k-th interval represents the segment where the k-th spline dominates the others.}

\item{highlight_band1}{a vector that states if a particular band of the plot has to be highlighted. It has to be a vector within the range \link{1,n_basis}}

\item{title_plot}{the title of the plot.}

\item{xtitle}{the title of the x-axis.}

\item{ytitle}{the title of the x-axis.}
}
\description{
This function gets the mean values of the regression coefficients as well as the lower and upper quantiles and builds the smooted curves with their credible bands.
It does not computes the quantiles nor the mean starting from all the sampled values, they have to be previously computed using Compute_QuantileBeta function.
}
