% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Compute_PosteriorMeans}
\alias{Compute_PosteriorMeans}
\title{Compute Posterior means of sampled values}
\usage{
Compute_PosteriorMeans(
  file_name,
  p,
  n,
  stored_iterG,
  stored_iter = 0L,
  Beta = FALSE,
  Mu = FALSE,
  TauEps = FALSE,
  Precision = TRUE
)
}
\description{
Compute Posterior means of sampled values
}
