% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Compute_QuantileBeta}
\alias{Compute_QuantileBeta}
\title{Generate Bspine basis and its derivatives}
\usage{
Compute_QuantileBeta(SaveBeta, lower_qtl = 0.05, upper_qtl = 0.95)
}
\description{
Generate Bspine basis and its derivatives
}
