% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Compute_Quantiles}
\alias{Compute_Quantiles}
\title{Compute quantiles of sampled values}
\usage{
Compute_Quantiles(
  file_name,
  p,
  n,
  stored_iterG = 0L,
  stored_iter = 0L,
  Beta = FALSE,
  Mu = FALSE,
  TauEps = FALSE,
  Precision = FALSE,
  prec_elem = 0L,
  lower_qtl = 0.05,
  upper_qtl = 0.95
)
}
\arguments{
\item{file_name, }{string with the name of the file to be open. It has to include the extension, usually \code{.h5}.}

\item{p}{integer, the dimension of the graph or the number of basis functions used. It depends on the type of output that is contained in \code{file_name}.
It has the same meaning of \code{p} parameter of \code{\link{GGM_sampling}}, \code{\link{FLM_sampling}} or \code{\link{FGM_sampling}}.}

\item{n}{integer, the number of observed data.}

\item{stored_iterG}{integer, the number of saved iterations for the graphical related quantities, i.e the graph and the precision matrix. Required only if \code{Precision} parameter is \code{TRUE}.}

\item{stored_iter}{integer, the number of saved iterations for the regression parameters, i.e \mjseqn{\beta}s, \mjseqn{\mu} and \mjseqn{\tau_{\epsilon}}.
Required if at least one of \code{Beta}, \code{Mu},  \code{TauEps} parameter is \code{TRUE}.}

\item{Beta}{boolean, set \code{TRUE} to compute the quantiles for all \mjseqn{p*n} \mjseqn{\beta} coefficients. It may require long time.}

\item{Mu}{boolean, set \code{TRUE} to compute the quantiles for all \mjseqn{p} parameters.}

\item{TauEps}{boolean, set \code{TRUE} to compute the quantiles of \mjseqn{\tau_{\epsilon}} parameter.}

\item{Precision}{boolean, set \code{TRUE} to compute the quantiles for all the elements of the precision matrix. Some care is requested.
If the file represents the output of \code{GGM}, \code{FGM} of \code{FLM} with fixed graph, the sampled precision matrices are of size \mjseqn{p \times p}. In that case set \code{prec_elem} parameter
equal to \mjseqn{\frac{p(p+1)}{2}}. Instead, for outputs coming from \code{FLM} sampler with diagonal graph, only the diagonal of the precision matrix is saved. If so, set \code{prec_ele} eqaul to \mjseqn{p}.}

\item{prec_elem}{integer, set equal to \mjseqn{p} if \code{file_name} represents the output of \code{FLM} sampler with diagonal graph. Set \mjseqn{\frac{p(p+1)}{2}} otherwise.}

\item{lower_qtl}{the level of the first desired quantile.}

\item{upper_qtl}{the level of the second desired quantile.}
}
\value{
It returns a list containig the upper and lower quantiles of the requested quantities.
}
\description{
\loadmathjax This function reads the sampled values saved in a binary file and computes the quantiles of the desired level.
}
