% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Create_RandomGraph}
\alias{Create_RandomGraph}
\title{Generates random Graphs}
\usage{
Create_RandomGraph(
  p,
  n_groups = 0L,
  form = "Complete",
  groups = NULL,
  sparsity = 0.5,
  seed = 0L
)
}
\arguments{
\item{p}{int, the dimension of the graph in its complete form.}

\item{n_groups}{int, the number of desired groups. Not used if form is \code{"Complete"} or if the groups are directly insered as group parameter.}

\item{form}{String, the only possibilities are \code{"Complete"} and \code{"Block"}.}

\item{groups}{List representing the groups of the block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. Leave \code{NULL} if the graph is not in block form.}

\item{sparsity}{double, the desired sparsity of the randomly generated graph. It has to be striclty positive and striclty less than one. It is set to 0.5 otherwise.}

\item{seed}{int, the value of the seed. Default value is 0 that implies that a random seed is drawn.}
}
\value{
the adjacency matrix of the randomly generated graph.
}
\description{
This function genrates random graph both in \code{"Complete"} or \code{"Block"} form
}
