% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{FLM_sampling}
\alias{FLM_sampling}
\title{Functional Linear model for smoothing}
\usage{
FLM_sampling(
  data,
  niter,
  burnin,
  thin,
  BaseMat,
  diagonal_graph = T,
  G = NULL,
  threshold_GWish = 1e-08
)
}
\arguments{
\item{data}{matrix of dimension r x n containing the evaluation of n functional data over a grid of r nodes.}

\item{niter}{the number of total iterations to be performed in the sampling. The number of saved iteration will be (niter - burnin)/thin.}

\item{burnin}{the number of discarded iterations.}

\item{thin}{the thining value, it means that only one out of thin itarations is saved.}

\item{BaseMat}{matrix of dimension r x p containing the evalutation of p Bspline basis over a grid of r nodes}

\item{diagonal_graph}{boolean, set true if the graph has to be diagonal. Set false otherwise and pass as input the desired graph through the G parameter.}

\item{threshold_GWish}{double, stop algorithm for GWishart samling if the difference between two subsequent iterations is less than threshold_conv. Used only if diagonal_graph is set to false.}
}
\value{
Two lists are returned, one with all the sampler values and the other with the posterior means.
}
\description{
This function performs a linear regression for functional data, according to model (INSERIRE FORMULA MODELLO).
It is not a graphical model, the graph has to be fixed. It is possible to fix both a diagonal graph or a generic graph.
}
