% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{FLM_sampling}
\alias{FLM_sampling}
\title{Functional Linear model for smoothing}
\usage{
FLM_sampling(
  data,
  niter,
  burnin,
  thin,
  BaseMat,
  diagonal_graph = T,
  G = NULL,
  threshold_GWish = 1e-08,
  progressBar = TRUE
)
}
\arguments{
\item{data}{matrix of dimension \code{n_grid_points x n} containing the evaluation of \code{n} functional data over a grid of \code{n_grid_points} nodes.}

\item{niter}{the number of total iterations to be performed in the sampling. The number of saved iteration will be \code{(niter - burnin)/thin}.}

\item{burnin}{the number of discarded iterations.}

\item{thin}{the thining value, it means that only one out of \code{thin} itarations is saved.}

\item{BaseMat}{matrix of dimension \code{n_grid_points x p} containing the evalutation of \code{p} Bspline basis over a grid of \code{n_grid_points} nodes.}

\item{diagonal_graph}{boolean, set true if the graph has to be diagonal. Set false otherwise and pass as input the desired graph through the \code{G} parameter.}

\item{G}{matrix of size \code{p x p} representing the graphical part of the model that would remain fixed through out the sampling. Needed only if \code{diagonal_graph} is set to \code{FALSE}.}

\item{threshold_GWish}{double, stop algorithm for GWishart samling if the difference between two subsequent iterations is less than \code{threshold_conv.} Used only if \code{diagonal_graph} is set to \code{FALSE}.}

\item{progressBar}{boolean, set \code{TRUE} to display the progress bar.}
}
\value{
Two lists are returned, one with all the sampler values and the other with the posterior means.
}
\description{
This function performs a linear regression for functional data, according to model (INSERIRE FORMULA MODELLO).
It is not a graphical model, the graph has to be fixed. It is possible to fix both a diagonal graph or a generic graph.
}
