% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GGM_sampling}
\alias{GGM_sampling}
\title{Testing all the Gaussian Graphical Models samplers}
\usage{
GGM_sampling(
  data,
  p,
  n,
  niter,
  burnin,
  thin,
  D,
  b = 3,
  MCprior = 100L,
  MCpost = 100L,
  threshold = 1e-08,
  form = "Complete",
  prior = "Uniform",
  algo = "MH",
  n_groups = 0L,
  seed = 0L,
  Gprior = 0.5,
  sigmaG = 0.1,
  paddrm = 0.5
)
}
\arguments{
\item{D}{Eigen Matrix of double stored columnwise. It has to be symmetric and positive definite.}

\item{b}{double, it is shape parameter. It has to be larger than 2 in order to have a well defined distribution.}

\item{MCiteration}{unsigned int, the number of iteration for the MonteCarlo approximation.}
}
\description{
Testing all the Gaussian Graphical Models samplers
}
