% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GGM_sampling}
\alias{GGM_sampling}
\title{Sampler for Guassian Graphical Models}
\usage{
GGM_sampling(
  data,
  n,
  niter = 1e+05,
  burnin = niter/2,
  thin = 1,
  Param = NULL,
  HyParam = NULL,
  form = "Complete",
  prior = "Uniform",
  algo = "RJ",
  groups = NULL,
  n_groups = 0,
  seed = 0,
  print_info = TRUE
)
}
\arguments{
\item{data}{matrix of size \eqn{p x p} containing \eqn{\sum(Y_i^{T}Y_i)}. Data are required to be zero mean.}

\item{n}{number of observed data.}

\item{niter}{number of total iterations to be performed in the sampling. The number of saved iteration will be \code{(niter - burnin)/thin}.}

\item{burnin}{number of discarded iterations.}

\item{thin}{thining value, it means that only one out of \code{thin} itarations is saved.}

\item{Param}{list containing parameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{sampler_parameters}} function. It is indeed recommended to build it through that particular function. \code{BaseMat} field is not needed.}

\item{HyParam}{list containing hyperparameters needed by the sampler. It has to follow the same notation of the one generated by \code{\link{GM_hyperparameters}} function. It is indeed recommended to build it through that particular function.}

\item{form}{string that may take as values only \code{"Complete"} of \code{"Block"} . It states if the algorithm has to run with \code{"Block"} or \code{"Complete"} graphs.}

\item{prior}{string with the desidered prior for the graph. Possibilities are \code{"Uniform"}, \code{"Bernoulli"} and for \code{"Block"} graphs only \code{"TruncatedBernoulli"} and \code{"TruncatedUniform"} are also available.}

\item{algo}{string with the desidered algorithm for sampling from a GGM. Possibilities are \code{"MH"}, \code{"RJ"} and \code{"DRJ"}.}

\item{groups}{a list representing the groups of the block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. If \code{"NULL"}, \code{"n_groups"} are automatically generated. Not needed if form is set to \code{"Complete"}.}

\item{n_groups}{int, number of desired groups. Not used if form is \code{"Complete"} or if the groups are directly insered as group parameter.}

\item{seed}{set 0 for random seed.}

\item{print_info}{boolean, if true progress bar and execution time are displayed.}
}
\value{
This function returns a list with the posterior precision mean, a matrix with the probability of inclusion of each link, the number of accepted moves, the number of visited graphs and the list of all visited graphs.
}
\description{
This function draws samples a posteriori from a Gaussian Graphical Models.
}
