% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{GM_hyperparameters}
\alias{GM_hyperparameters}
\title{Skeleton for Hyperparameters in Graphical Models}
\usage{
GM_hyperparameters(
  p,
  a_tau_eps = 20,
  b_tau_eps = 0.002,
  sigma_mu = 100,
  b_K = 3,
  D_K = NULL,
  p_addrm = 0.5,
  sigmaG = 0.5,
  Gprior = 0.5
)
}
\arguments{
\item{p}{integer, the dimension of the underlying graph.}

\item{a_tau_eps}{Shape parameter for gamma prior distribution of \code{tau_eps} parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{b_tau_eps}{Rate parameter for gamma prior distribution of \code{tau_eps} parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{sigma_mu}{covariance for normal multivariate distribution of \code{mu} parameter. Needed in Functional Models, i.e \code{FGM} and \code{FLM} samplers.}

\item{b_K}{prior GWishart Shape parameter. It has to be larger than 2 in order to have a well defined distribution. Needed in all graphical model, i.e \code{FGM} and \code{GGM} sampler and in \code{FLM} with fixed graph.}

\item{D_K}{matrix of size \mjseqn{p \times p}, it is prior GWishart Inverse-Scale parameter. It has to be symmetric and positive definite. Default is identity matrix. Needed in all graphical model, i.e \code{FGM} and \code{GGM} sampler and in \code{FLM} with fixed graph.}

\item{p_addrm}{the probability of proposing a new graph by adding one link. It has to be in the range \mjseqn{(0,1)}. Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler.}

\item{sigmaG}{the standard deviation used to perturb elements of precision matrix when constructing the new proposed matrix.
Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler. If algorithm is \code{"MH"} it is not used.}

\item{Gprior}{represents the prior probability of inclusion of each link in case \code{"Bernoulli"} prior is selected for the graph. It has to be in the range \mjseqn{(0,1)}.
Set 0.5 for \code{"Uniform"} prior. Needed in all graphical model where the graph is random, i.e \code{FGM} and \code{GGM} sampler.}
}
\value{
A list with all the hyperparameters described as possible inputs. All values that are not explicitely provided are defaulted.
}
\description{
\loadmathjax This function simply creates a skeleton for all the hyperparameters that has to be fixed in graphical samplers, that are \code{\link{GGM_sampling}} and \code{\link{FGM_sampling}}.
Prefer to use \code{\link{LM_hyperparameters}} for \code{\link{FLM_sampling}}.
All the quantities has a default value, moreover some of them may not be needed in some algorithm. The goal of this function is to fix a precise notation that will be used in all the code.
}
