% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_Gconstant}
\alias{log_Gconstant}
\title{The Monte Carlo method, developed by Atay-Kayis and Massam (2005), is implemented. It works with both decomposable and non decomposable graphs.
In particular it is possible to provide a graph in block form.}
\usage{
log_Gconstant(G, b, D, MCiteration = 500L, groups = NULL, seed = 0L)
}
\arguments{
\item{G}{matrix representing the desired graph. It has to be a \mjseqn{p \times p} matrix if the graph is in block form, i.e if groups is non null,
otherwise it has to be coherent with the number of groups. Only the upper triangular part is needed.}

\item{b}{GWishart Shape parameter. It has to be larger than 2 in order to have a well defined distribution.}

\item{D}{GWishart Inverse-Scale matrix. It has to be of size \mjseqn{p \times p}, symmetric and positive definite.}

\item{MCiteration}{the number of iterations for the Monte Carlo approximation.}

\item{groups}{a Rcpp list representing the groups of the block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. Leave \code{NULL} if the graph is not in block form.}

\item{seed}{integer, seeding value. Set 0 for random seed.}
}
\value{
log of the normalizing constant of GWishart distribution.
}
\description{
The Monte Carlo method, developed by Atay-Kayis and Massam (2005), is implemented. It works with both decomposable and non decomposable graphs.
In particular it is possible to provide a graph in block form.
}
