% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{plot_curves}
\alias{plot_curves}
\title{Plot curves}
\usage{
plot_curves(
  data1,
  data2 = NULL,
  range,
  n_plot = 1,
  grid_points = NULL,
  internal_knots = NULL,
  highlight_band1 = NULL,
  highlight_band2 = NULL,
  title_plot = "Curves",
  xtitle = " ",
  ytitle = " ",
  legend_name1 = "data1",
  legend_name2 = "data2"
)
}
\arguments{
\item{data1}{matrix of dimension \mjseqn{n\_curves \times n\_grid\_points} representing the first functional dataset to be plotted.}

\item{data2}{matrix of dimension \mjseqn{n\_curves \times n\_grid\_points} representing the second dataset to be plotted, if needed.}

\item{range}{the range where the curves has to be plotted. Not needed if \code{n_plot} is 0.}

\item{n_plot}{the number of curves to be plotted. Set 0 for no plot.}

\item{grid_points}{vector of size \mjseqn{n\_grid\_points} with the points where the splines are evaluated. If defaulted they are uniformly generated. Not needed if \code{n_plot} is 0.}

\item{internal_knots}{vector with the internal knots used to construct the splines. Can be obtained as return value of function \code{\link{Generate_Basis}}. Default is null.
If provided, \code{n_basis} are displayed in the plot. The \code{k}-th interval represents the segment where the \code{k}-th spline dominates the others.}

\item{highlight_band1}{a vector that states if a particular band of the plot has to be highlighted. It has to be a vector within the range \mjseqn{[1,n\_basis]}.}

\item{highlight_band2}{a vector that states if a particular band of the plot has to be highlighted. It has to be a vector within the range \mjseqn{[1,n\_basis]}.}

\item{title_plot}{the title of the plot.}

\item{xtitle}{the title of the x-axis.}

\item{ytitle}{the title of the x-axis.}

\item{legend_name1}{the name for \code{data1} to be printed in the legend.}

\item{legend_name2}{the name for \code{data2} to be printed in the legend. Used only is two datasets are actually plotted.}
}
\value{
No values are returned.
}
\description{
\loadmathjax This functions gets one or two dataset representig functional data and plot them. It does not smooth the curves, indeed it requires as input the data, not
the regression coefficients. Use \code{\link{smooth_curves}} function for that.
}
