% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read_InfoFile}
\alias{Read_InfoFile}
\title{Read information from file}
\usage{
Read_InfoFile(file_name)
}
\arguments{
\item{file_name, }{string with the name of te file to be open. It has to include the extension, usually \code{.h5}.}
}
\description{
Read from \code{file_name} some information that are needed to extract data from it.
}
