% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnormal}
\alias{rmvnormal}
\title{Sampler for Multivariate Normal random variables}
\usage{
rmvnormal(
  mean,
  Mat,
  isPrec = FALSE,
  isChol = FALSE,
  isUpper = FALSE,
  seed = 0L
)
}
\arguments{
\item{mean}{vector of size \code{p} representig the mean of the Gaussian distribution.}

\item{Mat}{matrix of size \mjseqn{p \times p} representinf the covariance or the precision matrix or their Cholesky decompositions.}

\item{isPrec}{boolean, set \code{TRUE} if Mat parameter is a precision, \code{FALSE} if it is a covariance.}

\item{isChol}{boolean, set \code{TRUE} if Mat parameter is a triangular matrix representig the Cholesky decomposition of the precision or covariance matrix.}

\item{isUpper}{boolean, used only if \code{isChol} is \code{TRUE}. Set \code{TRUE} if Mat is upper triangular, \code{FALSE} if lower.}

\item{seed}{integer, seeding value. Set 0 for random seed.}
}
\value{
It returns a \mjseqn{p}-dimensional vector with the sampled values.
}
\description{
\loadmathjax This function draws random samples from Multivariate Gaussian distribution. It allows for both covariance and precision parametrization.
It is also possible to pass directly the Cholesky decomposition if it is available before the call.
}
