% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwishart}
\alias{rwishart}
\title{Sampler for Wishart random variables}
\usage{
rwishart(b, D, isChol = FALSE, isUpper = FALSE)
}
\arguments{
\item{b}{double, it is shape parameter.}

\item{D}{matrix of size \eqn{p x p} representig the Inverse Scale parameter. It has to be symmetric and positive definite.}

\item{isChol}{boolean, set \code{TRUE} if Mat parameter is a triangular matrix representig the Cholesky decomposition of the precision or covariance}

\item{isUpper}{boolean, used only if isChol is \code{TRUE}. Set \code{TRUE} if Mat is upper triangular, \code{FALSE} if lower.}
}
\value{
It returns a p x p matrix.
}
\description{
This function draws random samples from Wishart distribution.
It is also possible to pass directly the Cholesky decomposition of the Inverse Scale matrix if it is available before the call.
}
