% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{sampler_parameters}
\alias{sampler_parameters}
\title{Skeleton for Parametes}
\usage{
sampler_parameters(
  MCprior = 500,
  MCpost = 750,
  BaseMat = NULL,
  threshold = 1e-14
)
}
\arguments{
\item{MCprior}{integer, the number of iteration for the MonteCarlo approximation of prior normalizing constant of GWishart distribution. Needed in \code{"MH"} and \code{"RJ"} algorithms.}

\item{MCpost}{integer, the number of iteration for the MonteCarlo approximation of posterior normalizing constant of GWishart distribution. Needed only in \code{"MH"} algorithms.}

\item{BaseMat}{matrix of dimension \mjseqn{n\_grid\_points \times n\_basis} containing the evaluation of all the \mjseqn{n\_basis} B-spine basis function in all the grid points.
May be defaulted as \code{NULL} but note this case it is not automatically generated.
Note that it is not needed by \code{\link{GGM_sampling}} but is mandatory in \code{\link{FGM_sampling}} and \code{\link{FLM_sampling}}. For those cases,
use \code{\link{Generate_Basis}} to generete it.}

\item{threshold}{threshold for convergence in GWishart sampler. It is not needed only in \code{FLM} sampler with diagonal graph.}
}
\value{
A list with all parameters described as possible inputs.
}
\description{
\loadmathjax This function simply creates a skeleton for some parameters that has to be fixed in all samplers. All the quantities has a default value,
moreover some of them may not be needed in some algorithm. The goal of this function is to fix a precise notation that will be used in all the code.
}
