% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGSL_export.R
\name{SimulateData_GGM}
\alias{SimulateData_GGM}
\title{Simulate data from GGM}
\usage{
SimulateData_GGM(
  p,
  n,
  n_groups = 0,
  form = "Complete",
  groups = NULL,
  adj_mat = NULL,
  seed = 0,
  mean_null = TRUE,
  sparsity = 0.3
)
}
\arguments{
\item{p}{dimension of the response variable that is the dimension of the underlying graph.}

\item{n}{number of observations.}

\item{n_groups}{number of desired groups. Not used if form is \code{"Complete"} or if the groups are directly insered as group parameter.}

\item{form}{string that states if the true graph has to be in \code{"Block"} or \code{"Complete"} form. Only possibilities are \code{"Complete"} and \code{"Block"}.}

\item{groups}{list representing the groups of block form. Numerations starts from 0 and vertrices has to be contiguous from group to group,
i.e ((0,1,2),(3,4)) is fine but ((1,2,3), (4,5)) and ((1,3,5), (2,4)) are not. Leave \code{NULL} if the graph is not in block form.}

\item{adj_mat}{matrix representing the desired graph. It has to be a (p x p) matrix if form is \code{"Complete"} and has to be coherent with the number of groups if form is \code{"Block"}.
If \code{NULL}, a random graph is generated.}

\item{seed}{seed used to simulate data. Set 0 for random seed.}

\item{mean_null}{boolean, set \code{TRUE} if data has to be zero mean.}

\item{sparsity}{desired sparsity in the graph. Not used if true graph is provided.}
}
\value{
a list containing a matrix called U with sum(Y_iY_i), the true precision, the true graph and if Block form is selected, the graph in its complete form is also returned.
}
\description{
Simulate data from GGM
}
